pyproject.toml
[project]
name = "kates"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "hdbscan>=0.8.40",
    "httpx>=0.28.1",
    "ipykernel>=6.30.1",
    "notebook>=7.4.7",
    "pandas>=2.3.2",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.20.0",
    "typer>=0.19.2",
    "umap-learn>=0.5.9.post2",
]
